% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppeldata.R
\docType{data}
\name{csp}
\alias{csp}
\title{Correlates of State Policy, version 2.1}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6018 rows and 2162 columns.}
\source{
\href{http://ippsr.msu.edu/public-policy/correlates-state-policy}{Correlates of State Policy}
}
\usage{
data(csp)
}
\description{
Website description:
}
\details{
> The Correlates of State Policy Project includes more than nine-hundred variables,
with observations across the U.S. 50 states and time (1900 – 2016). These variables
represent policy outputs or political, social, or economic factors that may influence
policy differences across the states. The codebook includes the variable name, a short
description of the variable, the variable time frame, a longer description of the variable,
and the variable source(s) and notes.

See the codebook here: \href{https://github.com/vladtarko/ppeldata/tree/master/codebooks}.
}
\examples{
data(csp)

}
\references{
Boehmke, Frederick J., and Paul Skinner. 2012. "State Policy Innovativeness Revisited."
    _State Politics and Policy Quarterly_ 12(3):303-29.

    Caughey, Devin, and Christopher Warshaw. 2015. "The Dynamics of State Policy Liberalism, 1936–2014."
    _American Journal of Political Science_ 60 (4): 899–913.

    Jordan, Marty P. and Matt Grossmann. 2017. The Correlates of State Policy Project v.2.1.
    East Lansing, MI: Institute for Public Policy and Social Research (IPPSR).
}
\keyword{datasets}
