% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppeldata.R
\docType{data}
\name{fraser}
\alias{fraser}
\title{Fraser Economic Freedom of the World}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3498 rows and 61 columns.}
\source{
\href{http://www.statepolicyindex.com/data/}{State Policy Database}
}
\usage{
data(fraser)
}
\description{
The components of the Fraser index of Economic Freedom.
}
\details{
The following variables are the five aggregated subcomponents:
Size_of_Government, Property_Rights, Sound_Money, Trade, Regulation.

The variables range from 0 to 10, with 10 describing highest economic freedom. E.g.
if Size_of_Government is closer to 10 it means the government involvement in the economy
is smaller; if Regulation is closer to 10 it means there are fewer regulations.

The `statepolicy_meta` contains the variable descriptions and original sources.
The `statepolicy` constains the data.
}
\examples{
data(fraser)

}
\references{
Jason Sorens, Fait Muedini and William P. Ruger. 2008. "U.S. State and Local Public Policies in 2006: A New Database." _State Politics and Policy Quarterly_
vol 8, no 3, pp 309-326.
}
\keyword{datasets}
