% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppeldata.R
\docType{data}
\name{manifesto}
\alias{manifesto}
\title{Manifesto Project Main Dataset of Party Preferences}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4456 rows and 175 columns.}
\source{
\href{https://manifesto-project.wzb.eu/datasets}{Manifesto Project}
}
\usage{
data(manifesto)
}
\description{
Cross-country panel data (1920-2018).
}
\details{
See the codebook here: \href{https://github.com/vladtarko/ppeldata/tree/master/codebooks}.
}
\examples{
data(manifesto)

}
\references{
Volkens, Andrea / Krause, Werner / Lehmann, Pola / Matthieß, Theres / Merz,
 Nicolas / Regel, Sven / Weßels, Bernhard(2019):
 The Manifesto Data Collection. Manifesto Project (MRG / CMP / MARPOR). Version 2019a.
 Berlin: Wissenschaftszentrum Berlin für Sozialforschung (WZB).
 \href{https://doi.org/10.25522/manifesto.mpds.2019a}

 Budge, Ian / Klingemann, Hans-Dieter / Volkens, Andrea / Bara, Judith with Tanenbaum,
 Eric / Fording, Richard C. / Hearl, Derek J. / Kim, Hee Min / McDonald, Michael / Mendez,
 Silvia (2001): Mapping Policy Preferences. Estimates for Parties, Electors, and Governments
 1945-1998. Oxford: Oxford University Press.

 Klingemann, Hans-Dieter / Volkens, Andrea / Bara, Judith / Budge, Ian / McDonald, Michael (2006):
 Mapping Policy Preferences II. Estimates for Parties, Electors, and Governments in Eastern Europe,
 the European Union and the OECD, 1990-2003. Oxford: Oxford University Press.
}
\keyword{datasets}
