% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppeldata.R
\docType{data}
\name{statepolicy}
\alias{statepolicy}
\title{United States State Policies}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4258 rows and 709 columns.}
\source{
\href{http://www.statepolicyindex.com/data/}{State Policy Database}
}
\usage{
data(statepolicy)
   data(statepolicy_meta)
}
\description{
The Ruger and Sorens Database of State and Local Public Policies and Policy Ideology Indices (2018)
}
\details{
The `statepolicy_meta` contains the variable descriptions and original sources.
The `statepolicy` constains the data.
}
\examples{
data(statepolicy)
data(statepolicy_meta)

}
\references{
Jason Sorens, Fait Muedini and William P. Ruger. 2008. "U.S. State and Local Public Policies in 2006: A New Database." _State Politics and Policy Quarterly_
vol 8, no 3, pp 309-326.
}
\keyword{datasets}
