% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppeldata.R
\docType{data}
\name{ted}
\alias{ted}
\title{Total Economy Database}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8750 rows and 19 columns.}
\source{
\href{https://www.conference-board.org/data/economydatabase/}{Total Economy}
}
\usage{
data(ted)
}
\description{
Cross-country panel data (1950-2019) on:
}
\details{
"Total GDP, in millions of 2018 US$"
"Persons employed (thousands)"
"Average annual hours worked per worker"
"Total annual hours worked (millions)"
"Midyear population (thousands)"
"Labor productivity per person employed in 2018 US$"
"Labor productivity per hour worked in 2018 US$"
"GDP per capita in 2018 US$"
"Growth of GDP, percent change"
"Growth of employment, percent change"
"Growth of total hours worked, percent change"
"Growth of population, percent change"
"Growth of Labor Productivity per person employed, percent change"
"Growth of Labor Productivity per hour worked, percent change"
"Growth of GDP per capita, percent change"

See the codebook here: \href{https://github.com/vladtarko/ppeldata/tree/master/codebooks}.
}
\examples{
data(ted)

}
\keyword{datasets}
