% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppeldata.R
\docType{data}
\name{vdem}
\alias{vdem}
\title{Varieties of Democracy full dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26834 rows and 3888 columns.}
\source{
\href{https://www.v-dem.net/en/data/data-version-9/}{Varieties of Democracy}
}
\usage{
data(vdem)
}
\description{
The full cross-section panel data: Country-Year: V-Dem Full+Others
All 450+ V-Dem indicators and indices + 59 other indicators from other data sources.
}
\details{
See the codebook here: \href{https://github.com/vladtarko/ppeldata/tree/master/codebooks}.
}
\examples{
data(vdem)

}
\references{
Coppedge, Michael, John Gerring, Carl Henrik Knutsen, Staffan I. Lindberg, Jan Teorell,
David Altman, Michael Bernhard, M. Steven Fish, Adam Glynn, Allen Hicken, Anna Lührmann,
Kyle L. Marquardt, Kelly McMann, Pamela Paxton, Daniel Pemstein, Brigitte Seim, Rachel Sigman,
Svend-Erik Skaaning, Jeffrey Staton, Steven Wilson, Agnes Cornell, Lisa Gastaldi, Haakon Gjerløw,
Nina Ilchenko, Joshua Krusell, Laura Maxwell, Valeriya Mechkova, Juraj Medzihorsky, Josefine Pernes,
Johannes von Römer, Natalia Stepanova, Aksel Sundström, Eitan Tzelgov, Yi-ting Wang, Tore Wig,
and Daniel Ziblatt. 2019. "V-Dem [Country-Year/Country-Date] Dataset v9",
_Varieties of Democracy (V-Dem) Project_. \href{https://doi.org/10.23696/vdemcy19}

Pemstein, Daniel, Kyle L. Marquardt, Eitan Tzelgov, Yi-ting Wang, Juraj Medzihorsky, Joshua Krusell,
Farhad Miri, and Johannes von Römer. 2019. “The V-Dem Measurement Model: Latent Variable Analysis
for Cross-National and Cross-Temporal Expert-Coded Data”, V-Dem Working Paper No. 21. 4th edition.
University of Gothenburg: Varieties of Democracy Institute.
}
\keyword{datasets}
