% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppeldata.R
\docType{data}
\name{qog}
\alias{qog}
\title{Quality of Government dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15403 rows and 2199 columns.}
\source{
\href{https://qog.pol.gu.se/data/datadownloads/qogstandarddata}{Quality of Government Institute}
}
\usage{
data(qog)
}
\description{
The full cross-section standard panel data.
}
\examples{
data(qog)

}
\references{
Teorell, Jan et al. 2018. The Quality of Government Standard Dataset,
   Version Jan18. University of Gothenburg: The Quality of Government Institute.
(\href{https://qog.pol.gu.se/data/datadownloads/qogstandarddata}{Quality of Government Institute})
}
\keyword{datasets}
